+++ src/modules/fsApi.js
@@ -1,27 +1,10 @@
-import { parse, join, isAbsolute, basename } from 'node:path';
-import {
-    stat,
-    readdir,
-    writeFile,
-    rename,
-    unlink,
-} from 'node:fs/promises';
-import { createReadStream, createWriteStream } from 'node:fs';
+import { parse, join } from 'node:path';
+import { stat, readdir } from 'node:fs/promises';
 
-const setRootPath = (path) => {
+const getRootPath = (path) => {
     try {
         const rootPath = parse(String(path)).root;
         return rootPath;
-    } catch (error) {
-        throw new Error(error);
-    }
-};
-
-const getFullPath = async (path, workingPath) => {
-    try {
-        const newPath = isAbsolute(path) ? path : join(workingPath, path);
-        await stat(newPath);
-        return newPath;
     } catch (error) {
         throw new Error(error);
     }
@@ -61,79 +44,4 @@ const getListFiles = async (path) => {
     }
 };
 
-const readFile = async (path, workingPath) => {
-    try {
-        const fullPath = await getFullPath(path, workingPath);
-        const readable = await createReadStream(fullPath);
-        return readable;
-    } catch (error) {
-        throw new Error(error);
-    }
-};
-
-const createFile = async (path, workingPath) => {
-    try {
-        const fullPath = join(workingPath, path);
-        await writeFile(fullPath, '', {
-            flag: 'wx',
-        });
-    } catch (error) {
-        throw new Error(error);
-    }
-};
-
-const renameFile = async (filePath, newFileName) => {
-    try {
-        await rename(filePath, newFileName);
-    } catch (error) {
-        throw new Error(error);
-    }
-};
-
-const copyFile = async (filePath, newDirectoryPath, workingPath) => {
-    try {
-        const fileFullPath = await getFullPath(filePath, workingPath);
-        const directoryFullPath = await getFullPath(
-            newDirectoryPath,
-            workingPath
-        );
-        const fileName = basename(fileFullPath);
-        const newFilePath = join(directoryFullPath, fileName);
-
-        const readable = createReadStream(fileFullPath);
-        const writable = createWriteStream(newFilePath);
-        readable.pipe(writable);
-    } catch (error) {
-        throw new Error(error);
-    }
-};
-
-const deleteFile = async (path, workingPath) => {
-    try {
-        const fileFullPath = await getFullPath(path, workingPath);
-        await unlink(fileFullPath);
-    } catch (error) {
-        throw new Error(error);
-    }
-};
-
-const moveFile = async (filePath, newDirectoryPath, workingPath) => {
-    try {
-        await copyFile(filePath, newDirectoryPath, workingPath);
-        await deleteFile(filePath, workingPath);
-    } catch (error) {
-        throw new Error(error);
-    }
-};
-
-export {
-    setRootPath,
-    getFullPath,
-    getListFiles,
-    readFile,
-    createFile,
-    renameFile,
-    copyFile,
-    deleteFile,
-    moveFile,
-};
+export { getRootPath, getListFiles };
